feat(converter): accept modern class file versions
Add a class file version policy that accepts supported modern majors through
Java 26 and rejects preview or future versions with clear errors.

Parse modern constant pool tags for dynamic constants, modules, and packages,
and skip class-level attributes by length so newer metadata does not break
basic class parsing.

+++ TotalCrossSDK/src/main/java/tc/tools/converter/java/JavaConstantInfo.java
@@ -1,19 +1,20 @@
 // Copyright (C) 2000-2013 SuperWaba Ltda.
-// Copyright (C) 2014-2020 TotalCross Global Mobile Platform Ltda.
+// Copyright (C) 2014-2021 TotalCross Global Mobile Platform Ltda.
+// Copyright (C) 2022-2026 Amalgam Solucoes em TI Ltda
 //
 // SPDX-License-Identifier: LGPL-2.1-only
 package tc.tools.converter.java;
 
 public final class JavaConstantInfo {
   public int index1, index2;
-  public byte type;
+  public int type;
 
-  public JavaConstantInfo(byte type, int index1) {
+  public JavaConstantInfo(int type, int index1) {
     this.index1 = index1;
     this.type = type;
   }
 
-  public JavaConstantInfo(byte type, int index1, int index2) {
+  public JavaConstantInfo(int type, int index1, int index2) {
     this.index1 = index1;
     this.index2 = index2;
     this.type = type;
